extract the DB-free queue readers so the claim-time guard is testable in CI
The sampler-repair contract test passed locally and failed in CI on
"Error: DATABASE_URL is missing". It imports assertQueuedArtPromptContract,
which lived in artJobQueueCoverage.ts, which imports prisma, which throws at
import time without a DATABASE_URL. contract-tests.yml is deliberately DB-free.
Locally a DATABASE_URL happened to be exported, so the dependency was invisible.

inferQueuedArtEngine, queuedArtSamplerSettings, queuedArtPrompt and
assertQueuedArtPromptContract need no database — they read a payload. They now
live in server/utils/artJobQueueSettings.ts; artJobQueueCoverage re-exports all
three of its public ones, so every existing caller keeps its import path and
verifyFacetCatalogMaintenance's textual guard still finds them.

The test asserts the separation holds: artJobQueueSettings must not import
prisma or artJobQueueCoverage. Moving them back would break no runtime caller —
which is exactly why it would go unnoticed. It would only make the contract
test unrunnable in CI, silently, the way it already was once.

Verified: verifyArtJobSamplerRepair with DATABASE_URL unset (the CI condition)
passes; verifyFacetCatalogMaintenance, verifyArtPromptContract and
verifyArtPromptRepair pass; vue-tsc --noEmit exits 0; eslint clean on every
touched file and the lint ratchet is unchanged at 392.
